requirements.txt
Django==3.2.7
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-filter==21.1
django-session-timeout==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
python-dotenv==0.19.1
redis==3.5.3
psycopg2-binary==2.9.1
channels==3.0.4
pydantic==1.8.2